#194 Button Component 리펙토링
