--- .storybook/addons.js
@@ -1,1 +1,0 @@
-import "storybook-addon-styled-component-theme/dist/register";
