+++ src/components/atoms/button/style.less
@@ -1,0 +1,59 @@
+@button-orange      : #f2711cd9;
+@button-light       : #fff;
+@button-light-border: #a7a9af;
+@button-dark        : #444a4c;
+@button-dark-border : #444a4c;
+
+.csms {
+    border-radius: 3px;
+
+    &.sub-default {
+        color       : #fff;
+        background  : @button-orange;
+        border-color: @button-orange;
+
+        &:hover {
+            color       : #fff;
+            background  : darken(@button-orange, 10%);
+            border-color: darken(@button-orange, 10%);
+        }
+    }
+
+    &.sub-light {
+        color: #222;
+        background  : @button-light;
+        border-color: @button-light-border;
+
+        &:hover {
+            color: #222;
+            background  : darken(@button-light, 6%);
+            border-color: darken(@button-light-border, 8%);
+        }
+
+        &:focus,
+        &:active {
+            color: #222;
+            background  : darken(@button-light, 12%);
+            border-color: darken(@button-light-border, 8%);
+        }
+    }
+
+    &.sub-dark {
+        color       : #fff;
+        background  : @button-dark;
+        border-color: @button-dark-border;
+
+        &:hover {
+            color       : #fff;
+            background  : lighten(@button-dark, 10%);
+            border-color: lighten(@button-dark-border, 10%);
+        }
+
+        &:focus,
+        &:active {
+            color       : #fff;
+            background  : darken(@button-dark, 10%);
+            border-color: darken(@button-dark-border, 10%);
+        }
+    }
+}
